Accommodate use cases where capacity exceeds size_t::max()
size_t is guaranteed to be large enough for any _single_ in-memory
object size, but ClpMap may accumulate multiple large objects.

Technically, if we remove sizeof() components in MemoryCountedFor(),
then we could even use ClpMap for managing on-disk caches (where the
capacity is even more likely to exceed size_t::max()).

+++ src/ssl/context_storage.h
@@ -46,7 +46,7 @@ public:
     virtual bool aggregatable() const { return false; }
 };
 
-inline size_t MemoryUsedByContext(const Security::ContextPointer &) {
+inline uint64_t MemoryUsedByContext(const Security::ContextPointer &) {
     return 1024; // TODO: Calculate approximate memory usage by the context.
 }
 using LocalContextStorage = ClpMap<SBuf, Security::ContextPointer, MemoryUsedByContext>;
